:truck:refactor: ajuste na estrutura de arquivos e pastas
