+++ .env.example
@@ -3,3 +3,5 @@ DB_PORT=
 DB_USER=
 DB_PASSWORD=
 DB_NAME=
+DB_DIALECT=sqlite
+DB_STORAGE=:memory:
