+++ config/database.js
@@ -1,0 +1,11 @@
+require('dotenv').config();
+
+module.exports = {
+  "username": process.env.DB_USER,
+  "password": process.env.DB_PASSWORD,
+  "database": process.env.DB_NAME,
+  "host": process.env.DB_HOST,
+  "port": process.env.DB_PORT,
+  "dialect": process.env.DB_DIALECT || "mysql",
+  "storage": process.env.DB_PORT
+};
