+++ index.js
@@ -3,8 +3,8 @@ const app = express();
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
 const connection = require('./database/database');
-const Pergunta = require('./database/Pergunta.js');
-const Resposta = require('./database/Resposta.js');
+const Pergunta = require('./model/Pergunta.js');
+const Resposta = require('./model/Resposta.js');
 const session = require('express-session');
 const flash = require('connect-flash');
 require('dotenv/config');
