+++ model/Resposta.js
@@ -1,5 +1,5 @@
 const Sequelize = require('sequelize');
-const connection = require('./database');
+const connection = require('../database/database');
 
 const Resposta = connection.define('Resposta', {
 	corpo: {
