+++ package.json
@@ -19,6 +19,7 @@
     "express": "^4.17.3",
     "express-session": "^1.17.2",
     "mysql2": "^2.3.3",
-    "sequelize": "^6.19.0"
+    "sequelize": "^6.19.0",
+    "sqlite3": "^5.0.6"
   }
 }
